scripts/data_processing/requirements.txt
datasets
openai
pinecone-client
numpy
tqdm
python-dotenv
aiohttp
httpx